UI CHANGE +  NOTI

+++ lib/core/widgets/animated_counter.dart
@@ -49,7 +49,7 @@ class _AnimatedCounterState extends State<AnimatedCounter>
     if (oldWidget.value != widget.value) {
       _previousValue = _animation.value;
       _animation = Tween<double>(
-        begin: _previousValue.toDouble(),
+        begin: _previousValue,
         end: widget.value.toDouble(),
       ).animate(CurvedAnimation(
         parent: _controller,
